+++ src/main/java/kg/megalab/kindergarten/models/dto/GroupCategoryDto.java
@@ -15,5 +15,5 @@ public class GroupCategoryDto {
     Long id;
     String name;
     boolean active;
-    double price;
+    Integer price;
 }
